SQLDatabaseToAWS/requirements.txt
# ── AWS ───────────────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0

# ── Database ──────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0

# ── Linting / Formatting ──────────────────────────────────────
ruff>=0.4.0

# ── Data seeding (scripts/seed_data.py) ───────────────────────
Faker>=24.0.0
tqdm>=4.66.0

# ── ETL / Glue jobs ───────────────────────────────────────────
pandas>=2.2.0
pyarrow>=15.0.0

# ── Utilities ─────────────────────────────────────────────────
tabulate>=0.9.0
configparser>=6.0.0
